all non endpoint functions have been removed from washee_box_entry
